requirements.txt
flask
flask_mail
requests
python-dotenv







